requirements.txt
black
flake8
os
cv2
flask
spotipy
python-dotenv
functools
cvzone
